Update patch set 3
Patch Set 3: Code-Review+1 Verified+1

confirmed in master.

There was no mechanism in init to support an unset property (eg: logd.manual) so we opted for 'always on' for this feature in the logd source code rather than runtime property switch in logd and in init.rc. The developer will have to commit a rootdir/init.rc with logd disabled in order to use the feature. We fortified the error message to incorporate some selinux enforcement adjustments to resolve an issue with some clients lacking the necessary permissions to log to the u:r:su:s0 target by using runcon appropriately.

Patch-set: 3