Update patch set 2
Patch Set 2:

Regarding msg=, after this is merged I was hoping to get audit rules support merged in, which would allow records that are not avc denials to be emitted by the audit framework.

We might be able to infer what kind of record it is in some cases but it is inconvenient when the information is readily available.

Patch-set: 2