Update patch set 1
Patch Set 1:

Currently, we use this patch so that we can perform some actions when a USB Ethernet adapter is connected, basically configure and bring up the interface. We want to support devices that don't have OTG hardware for ADB & Fastboot. Without a change like this, the Ethernet adapter has to be connected at boot and never disconnected.

Patch-set: 1