Update patch set 1
Patch Set 1:

Another use-case is that there is also some sensor hub hardware that disappears and reappears when the device goes in and out of suspend and has to be re-configured when it comes back. Srinivas would know more about this one.

Patch-set: 1
Reviewer: Andrew P Boie <1011333@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0