Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

If you set the logd.auditd property to a value of 0, this will allow us to bisect if this new logging feature is the cause of the problem(s). Of note, the "logd.writer" thread is for all purposes regardless of the number of threads feeding the logger with their ALOG* messages, single threaded, much like "logd.auditd"; but their combination will result in multiple threads feeding the internal log database.

The log database has a mutex lock, so it should be safe, this CL allows us to explore the FUD that it is not ...

Patch-set: 1
Label: Code-Review=+1
Label: Verified=+1