Update patch set 2
Patch Set 2:

Yes, a bug! tv_sec is hard-coded to a u32 (it did not used to be that way, the fix to enforce that should have also changed LONG_MAX to U32_MAX (to be locally defined).

But, how did you get the warning from compilation and I did not(!) (my tests are hammerhead, volantis64 on goog/master and aosp_(arch)-eng on aosp/master).

Patch-set: 2