Update patch set 1
Patch Set 1:

Let me restate, in the form of a plan

Correct sys/cdefs.h in bionic to incorporate the #ifdef here to __printflike as the 'best example' for all platforms as enh has suggested.

In android/log.h do an #ifndef __printflike for a local replacement and use that macro instead of __attribute__ directly.

Once a sys/cdefs.h is available on all compilation platforms, the #ifdef in android/log.h can be deprecated.

Patch-set: 1