Update patch set 2
Patch Set 2:

I don't think you can use TARGET_CPU_ABI. That variable will always be the name of the "primary" ABI of the given device. It doesn't get changed back and forth (remember that the file is parsed only once).

You could however do something godawful like

   TARGET_CFLAGS_64 += -DTARGET_ARCH=$(TARGET_CPU_ABI)
   TARGET_CFLAGS_32 += -DTARGET_ARCH=$(TARGET_2ND_CPU_ABI)

This will assume that TARGET_CPU_ABI is always 64 bit, but this wouldn't be the only place that assumes that.

Patch-set: 2