Update patch set 1
Patch Set 1:

To play devil's advocate: Our message regarding SELinux on Android up to this point so far has been that of adding an additional layer of security, whereas here we are actually removing the previous security mechanism.  I don't think modifying property_service.c is such a bad thing when handling properties, whereas relying solely on SELinux policy could lead to some issues as it is currently poorly understood in the general developer population, very easily changed, and would introduce a single point-of-failure.  Of course, these could easily be arguments for this change as well: a centralized policy is easier to reason about and protect, but I'm wondering what benefits we get from removing it at this point.

Patch-set: 1