Update patch set 1
Patch Set 1:

I suspect the situation you're describing is rare. 

One option is to just assume the risk that this will happen, and make sure we let people know that they shouldn't install untrustworthy programs where setenforce=0. We could, for example, modify the output of "setenforce" to print a warning message if the user disables SELinux.

Another option is to always use the SELinux rules unless SELinux is disabled. If SELinux is disabled, then only use the old rules. We'd maintain both sets of rules, and switch between them based on the SELinux enabled status. The risk is that, over time, the old rules will become stale. People will have different security rules in place depending on whether or not SELinux is enabled, which could be confusing.

Colin: do you have any suggestions on how to resolve your concerns?

Patch-set: 1