Update patch set 1
Patch Set 1:

My concern is basically the same as Daniel's - we've always said that selinux is in addition to the existing security, which makes sense if there is a big hammer to turn it off (setenforce 0).  We have to be careful moving security entirely into selinux.  But if you're ok with it, them I'm ok with it.

Patch-set: 1