Update patch set 1
Patch Set 1:

To clarify as this was discussed out-of-band from this CL (and because 'without regards to UID' is a barely acceptable simplification, but is not pedantically correct, sorry Nick). I figure recording here is a good way as any of understanding the history of the reason for this CL.

'eng' build, adb shell (uid UID_ROOT and gid UID_ROOT), then su to shell user. Credentials are pure shell (UID and GID are 2000 or UID_SHELL). At least before this CL ...

'userdebug' build, so shell spawned as uid UID_SHELL, and gid UID_SHELL, UID_LOG and UID_SYSTEM.

The shells are 'different', yet the intention is, on eng builds, that an su to shell should give you the same as the native shell in userdebug builds.

One example of where this difference creates a problem was with logd:

eng build -> adb shell -> su shell -> logcat -c (clear logs) -> (cred socket) logd. logd senses the difference and clears the logs associated only for UID_SHELL (2000); the expectation is that this 'path' should result in enough credentials to clear *all* the logs. With the change, logd will see the wire credentials including the gid of UID_LOG or UID_SYSTEM or UID_ROOT and grant it the extended privileges.

This CL ensures correct credentials via su. Thanks to Nick for recognizing this problem and solving it cleanly like this.

Patch-set: 1