Update patch set 1
Patch Set 1:

This was a result of the 'big C++ rewrite' several moons ago colliding with my new feature of splitting the lines in the tombstone files. A conflict resolution at the hands of others to clean up the mess resulting from multiple patches going in wibbly-wobbly-temporality resulted in this fragment of code without the line-wrap. *This* conflict resolution should finally clean up this locality of discrepancy :-)

Patch-set: 1