Update patch set 2
Patch Set 2:

(1 comment)

I don't see any code in logd which is conditional on AUDITD_LOG_TAG. Just to be clear, these logs are coming from userspace, not the kernel. Based on my (admittedly incomplete) understanding of logd, the only place where we do double writing of log entries is when the log entry comes from the kernel. Not from userspace.

Patch-set: 2