Update patch set 2
Patch Set 2: Code-Review+2

(1 comment)

Accepting deferral of endianness, but it will bite us one day ...

We assume we have AUDITD_LOG_TAG define, for if not, it is an error and we want the build to stop dead.

we need to clarify double ...

*Triple* writing is done:
1) main log id so that the application/tombstone/logcat-default-output/developers will easily see the violations in-line with their applications operations.
2) events log id were for your collection so that you could mine for violations.
3) kmsg log ... this is only so that the kernel-space message were also present in the traditional dmesg location.

If coming from userspace, I see no reason why they should not be in both 'main' and 'events' as well duplicated; but the dmesg logs rightfully should be left alone (?) for these class of violations.

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2