Update patch set 2
Patch Set 2:

> I'm fine with this if you want it, but it seems out of place.  If
 > netd can survive a stop/start after your other change, why do you
 > want to stop and start it here?  I believe the onrestart in init.rc
 > is to deal with cases where the native daemon has gotten into a bad
 > state and is what is causing the runtime restart, not to tie the
 > lifecycles of zygote and netd together.

Ok, I thought that _was_ the point of onrestart, to tie the lifecycles together, as you put it. Indeed, that's what's needed, and that's what this CL tries to maintain.

The framework (via system_server) contains a bunch of state (such as which "netIds" are valid) which is shared with netd. If the system_server is restarted, but netd is not, their states get out of sync. This manifests usually as the framework trying to create what it thinks is a new netId, but repeatedly getting rejected by netd because it's "already in use".

The other CL you referred to merely makes it so that state maintained by the kernel (in the form of RPDB rules) is flushed when netd is restarted. Doesn't affect the requirement for netd to in fact be restarted when the system_server is.

Patch-set: 2