Update patch set 1
Patch Set 1:

> This will do weird things to the progress bar if somebody locally
 > removes or moves the file while it is being downloaded.  What about
 > adding adb_fstat that can translate back to the windows file
 > descriptor if necessary?

The progress will just stop displaying if the file gets moved or renamed.  The only weird case I can think of is if someone overwrites the file that is being downloaded.

In any case, it seems like the file transfer should behave exactly the same regardless of if progress is displayed or not.  Right now, turning on the progress can cause transfer failures that might not have failed if the progress were off.

Patch-set: 1