Update patch set 2
Patch Set 2:

(2 comments)

I agree about reliability. If we have 100% confidence, *only* then will we remove the 30s timeout on non-blocking requests on the reader code in liblog.

Right now, the code shaves off 30 seconds of wonderment if logd should crash. Please note that *currently* something is preventing logd from producing tombstones when signalled, so this code is meaningless until we get to the root cause of that regression.

Patch-set: 2
Label: Verified=0