Update patch set 1
Patch Set 1:

pthread_exit() in some implementations is a longjmp to a protected exit handler with cleanup operations. The cleanup operations are performed in exit with the stack cleanup up. With the terminate without join flags, still protected, the stack is cleaned up and returned atomically with the thread termination.

pthread_cleanup_pop(true) is performed with the full stack, yes protected typically, but not exited, returning back to the calling thread unprotected _then_ a return or pthread_exit would be called once again placing us in the protected pthread library code, the cleanup has already been popped off.

The difference is that another thread can execute between cleanup and full thread exit in some implementations. For the previous case of us spawning transitory threads, can be deadly if a new replacement thread is created. For this case with the adjustments for persistent threads, _meh_

Patch-set: 1