Update patch set 3
Patch Set 3:

> It's just a result of our decision to trade support for archives
 > with multiple encodings for performance. Ideally we would like all
 > archives to be UTF-8 only. The ones which use only IBM PC are fine
 > too. The one which mix encodings but use only English are fine too.
 > 

This seems to trade correctness for performance... (I mean not this CL but the handling of multi-encodings).

IMO we should would add case an enum/int saying which type of encoding is used (i.e. mixed/utf8/imbpc etc). Then leave it to the user how he want to proceed. I think it would make stuff clearer. As it is now we may claim the encoding is utf8 even when it's a mixed one which might lead to wrong assumptions about correctness.

Patch-set: 3
Label: Verified=0