Update patch set 1
Patch Set 1:

> A question, on a native ISA environment, as the ro.dalvik.vm.native.bridge
 > is empty, art::Runtime::native_bridge_library_path_ will be empty
 > too. So SetupNativeBridge() will not be called and the art runtime
 > will not initialize libnativebridge. So the user can still install
 > a user side native-bridge from their app lib folder, because the
 > app lib folder has already been updated to linker by ART runtime.
 > Right?

Valid concern. However, as discussed in the mails this can't really happen as the app can't spawn another vm (and even if it could it would be in it's own process space so it wouldn't matter).

That said, thanks for drawing attention! It's a good spot and I think we should always be sure we setup nb. I will update the art code so that it does so.

Patch-set: 1