Update patch set 1
Patch Set 1:

> The tests in question are all kernel tests only?
I don't think so, I think other tests that need the root permission should have the same problem.
 > And you can use the same kernel for your -user and -userdebug
 > builds, right?
yes, but actually we want to provide only the user build,
and only test on the user build.
 > So you could run your kernel tests that require root on a
 > -userdebug build using su or adb root, and run the CTS on -user
 > build, and still know that the kernel test result carries over to
 > your -user build?
I am not sure if I understand you correctly here. We want to run all the existing test on user build, and hope to have the same result as run on eng build.

 > Wasn't it already the case that CTS required a user build, even
 > before the SELinux tests were introduced? 
That's true.
 > How did this work even
 > before SELinux was in the picture?  You already had to omit/exclude
 > certain CTS tests if running on non-user build, and you couldn't
 > run your tests as root on a user build.  Not seeing how this is new
 > or unique to SELinux.
Yes, this problem is the same as before.
Before we release the eng build and run all the tests on eng build. CTS test did have failures that should be passed if tested on user build.


Now we want to move to the user build from eng, and met this problem that not all tests can be run on user build, and especially the user build needs selinux enabled, which add more limits.

So I am not sure if we should provide some framework for tests even on the user build, so that developers only need to do the tests on user build for one time, instead of doing test twice or more on different builds.
Like here I want to use the ro.debuggable to control the console service, and disable the console service when release. if the console service is disabled, there is still security problem?
not like the eng build and user build we have now.
the eng/userdebug/user builds seem to be used for this purpose, but they are different builds, and seems have much difference there.

Patch-set: 1