Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

I don't like that the positive bind-mount test is going away.

I understand why it's now harder to do this. But this was the only test that tested the bind-mount (OK, given, you have to do it by hand to run with sudo, but at least it's there). The proper way to do this would be to have a fake .so for the test, just like the ART test is doing.

For now, please refactor the bind-mount code out and make it testable, similar to the name testing.

Patch-set: 6
Label: Code-Review=-1