Update patch set 1
Patch Set 1:

(8 comments)

So, deciding between __MINGW__ and _WIN32 will probably vary. For HAVE_FORKEXEC, it will always be _WIN32 because fork/exec is an OS thing, and will not vary between mingw and clang (which we will eventually move to). The one I said should be __MINGW__ looked like a library thing to me, which I believe was specific to __MINGW__.

Patch-set: 1
Label: Code-Review=0