Update patch set 1
Patch Set 1:

We expect the bootloader to see the "power off" reboot and decide whether to boot to charger mode or power down.  This catches the case in which the charger is disconnected prior to the bootloader being entered and choosing charger mode (which will shut down if the charger is removed after that point).  I don't believe it is expected that the PMIC have a special function to boot to charger-mode, only that the bootloader interpret this.

Patch-set: 1
Label: Code-Review=0