Update patch set 1
Patch Set 1:

it doesn't look like it would be much work to just kill TARGET_CPU_SMP, and that's a much safer (and more forward-looking) fix. right now you have to translate TARGET_CPU_SMP to ANDROID_SMP in every [target in every] .mk file, and some of the non-SMP cruft we've already removed has been cases where the .mk file was missing that magic. the current system is inherently unsafe, and no one is actually using it on purpose.

Patch-set: 1