Update patch set 2
Patch Set 2:

The problem in both adb and fastboot is a result of the way the apps try to get the path of their executable. The existing implementation uses GetCurrentProcess and ProcessInformationCopyDictionary which were deprecated in the 10.9 SDK. I'll code up an alternative which was introduced in 10.6 and seems to be the current recommended replacement implementation.

Patch-set: 2
Reviewer: Al Sutton <1010485@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0