Update patch set 1
Patch Set 1:

yeah, sorry --- too many changes in flight. for some reason i thought this and the matching toybox change had already been approved yesterday, but as far as i can tell there _was_ no matching change.

that said, these changes are not really reviewable. you have to trust that i've done a good enough job of ensuring compatibility between the old and new command.

on the bright side toybox does have tests, so if there's behavior we need to rely on in future we can write a test. (of course, we're not actually _running_ those tests. they don't all pass on the host either.)

Patch-set: 1