Update patch set 2
Patch Set 2:

Hi, Stephen

the change here use the directory of /data/bootchart for bootchart to read or write files,
 I checked it with eng build, which does not have much problem, should have problem when in user mode.
 But before I check with the user mode, if you have any suggestions please let me know.

And here is the log when run in the eng mode with nexus10, 
but there are some avc denial message, do you know why they are caused 
and how can we fix it? the dmesg log is here:
http://pastebin.com/H8qJrnhM

Thanks in advance!

Patch-set: 2