Update patch set 1
Patch Set 1:

enh, we're almost done with the cleanup (bluetooth and chromium are the last things, and chromium will be fixed by torne tomorrow, with local things I can actually do a checkbuild for N5 now). So can we use this as a bridge until you get around deleting the things you want to delete?

Patch-set: 1