Update patch set 1
Patch Set 1:

(2 comments)

> ... so even if your target is flounder you'll build manta audio
 > code and so on.

checkbuild scans the entire tree and builds every module it finds, but that's actually somewhat misleading, because if a module is hidden behind a `if ($(TARGET_PRODUCT),manta)` or whatever, it won't be "found" for other targets. I'm fairly certain target specific drivers and the like are guarded in this way (though I may have the variable name wrong), since it's always the drivers for other targets that bite me in the ass.

> if we ask bionicbb to retry on a change it wouldn't otherwise
> have tried, will it notice? or is it only listening in "our"
> directories?

It builds anything it sees an email for. While it won't proactively check projects outside the few I have project watches on, it will build your change if you add it as a reviewer (which I did before realizing that you had already submitted this).

Patch-set: 1