Update patch set 1
Patch Set 1:

but which changes _shouldn't_ have a version bump in that world? why not the USB3 changes? or danalbert's work to remove the duplication? or...?

i agree that the current "adb version" is not very useful, but i don't think this addresses the problem.

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0