Update patch set 1
Patch Set 1: Code-Review-1

This does not pass security muster within the application in liblog and to an uncontrolled target where PII could be exposed. Vendors are welcome to use such a change for their private debugging I expect.

We are exploring other alternatives such as implementing klogd (so that kernel messages are mixed with Android logger messaged), debug syslogd relay (to allow messages to propagate off the device to a puppetmaster into a lab setting, or to dedicated hardware such as noted here) and pstore (for kernel crash android logs recovery in bugreports) that could meet the requirements in a more upstream linux manner. Securing PII with MAC and DAC at every phase.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1