Update patch set 5
Patch Set 5:

Just a thought, would it be better to just use the output of stat to figure out which setfilecon() to call? This way it just works, and then we dont even need to handle options, or is their some nice property of making the user specify the type?

Patch-set: 5
Reviewer: William C Roberts <1013433@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0