Update patch set 1
Patch Set 1:

If a platform is dropping logs, it is recommended to adjust /proc/sys/net/unix/max_dgram_qlen so that they do not happen. The value can be made as large as necessary but will consume kernel memory when pushed.

We currently set it to 300 as this represents a form of logspam rate limiting.

Knowing if_your platform suffers from lost logs can be determined at the sender's side by checking for -EBUSY return code from the ALOG* handlers. If logs _must_ get to the destination then recommend a retry loop on that error code.

Patch-set: 1