Update patch set 1
Patch Set 1:

Totally up to you if you want this or not; just thought we'd share.  Checking at the client side isn't practical in our case.  No matter what values we set for /proc/sys/net/unix/max_dgram_qlen we were getting copious numbers of drops due to the inefficiencies in logd when there's a persistent log reader, which we use when Dogfooding.

If https://android-review.googlesource.com/124661 is accepted, the drops that actually occur drop to nearly 0 in all cases, and the need for this is greatly reduced.

Patch-set: 1