Update patch set 2
Patch Set 2:

It impacts when the function is called.

A stub function allows it to always be called in the init source (and optimized by compiler if no overrides).  If it is 0, we need a NULL check before calling it.  Thought it will be cleaner in the base init source.

Patch-set: 2