Update patch set 4
Patch Set 4:

Most comments are the performance reports ;-> which are for your benefit. I have incorporated the changes as comments or corrected code thus far except for the Patch Set 4 Pass which I am currently evaluating and requesting comments for.

For your edification: BM_log_print_overhead_fifo is a real-world(tm) overhead at less than 50% CPU utilization (screen off, CPU clocks are lowish, kind of like when audio is playing). BM_log_maximum_fifo is what happens at above 90% CPU utilization but likely much of your content is being dropped on the floor. BM_log_latency_fifo is the effective accuracy of your timestamp. BM_log_delay_fifo is how long it takes for a log message to show up in a logcat report. BM_log_delay_fifo - BM_log_delay is roughly the time it takes for the background thread to fire up and process the log message and hand it off to the logger.

SCHED_BATCH (which is the second lowest ahead of SCHED_IDLE) in the background thread. The caller has its own scheduling policy and can the more realtime-like SCHED_FIFO or SCHED_RR if need be.

getuid() is a necessary component of Android logging. The AID_ROOT is merely an initial value for a cache for the call results since it incurs a syscall. There are no AID_ROOT processes, except on incept when spawned by init. It is a requirement that the process quickly transition to an appropriate AID other than root for the UID.

Patch-set: 4