Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

John, it is more than compatibility. My first implementation of logcat in the face of the new logger had this format. The triage owners _only_ wanted to see when each log buffer started so they could see the overlap of their durations and did not care for the noise of log breaks. In fact, it was listed as a regression as it had happened before. The style has human implications, when someone asks for multiple log buffers, they want to see them mix. -D is in support of scripting, a small price to pay.

Patch-set: 2
Label: Code-Review=+1
Label: Verified=+1