Update patch set 1
Patch Set 1:

"Hardware" goes away on ARM64 platforms starting with this commit:

https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/arch/arm64/kernel/setup.c?id=44b82b7700d05a52cd983799d3ecde1a976b3bed

which will appear in upstream 3.19, and which we've backported to 3.14 since it's bundled together with other /proc/cpuinfo changes we want.  (This replaces our Android-specific /proc/cpuinfo hacks in 3.10.)

Devices can pass androidboot.hardware on the kernel command-line as an alternative.  In fact on Intel + MIPS this is the *only* way init can get this information since "Hardware" is an ARM-specific field.  Moving forward we just need to make sure we don't break this option.

Patch-set: 1