Update patch set 1
Patch Set 1:

This is really interesting.

The implementation lacks ability to expand properties passed on the commandline, which probably is useful. Also there will be problems for the exec to interact with android system properties. And we cannot set the security context, which may be useful for binaries on a rootfs.

Patch-set: 1
Label: Verified=0