Update patch set 2
Patch Set 2:

FYI: To _measure_ logging data loss, we are adding https://android-review.googlesource.com/#/c/132100. This should help discover and annotate where the data may be missing to either help triage, or max_dgram_qlen tuning.

Thanks for your feedback.

I appreciate that your design of the environment variable is used to try to suppress data loss. However, I have discovered that data rate is fungible, and may even be harder to characterize since https://android-review.googlesource.com/#/c/128571. logd can be delayed from doing its job if the collection of background processes consume more than a limited percentage of the total. This is by desire and design. As a result, rate limiting may not guarantee delivery, checking for EAGAIN errno from the write call is the only way. Also, zero impact in most cases; except if one application does this, others will necessarily suffer when the pipe is stuffed full ;-}

The goal will be that we will need to set max_dgram_qlen to a value that effectively reduces the losses to zero "for standard and normal logging activities"(tm). Fear of losing logging data should never IMHO result in a reaction with code like this; FUD must be zero at all times or logging is useless(tm) ;-/

However, if you respin this CL to instead check for the far more caustic LIBLOG_GUARANTEE_DELIVERY (or a better substitute name you come up with) to retry on EAGAIN internally with _strong_ comments in the library code _and_ in the associated README that the impact will result in the chattiest callers reduced to effectively background priority at logging points with the associated Heisenberg effects. And that other applications may be denied logging services. A disclaimer, for reasons of design purity, should also state that if one suspects data loss under normal logging load, they should instead focus on tuning max_dgram_qlen for the platform. The environment variable is for temporary mitigation of applications that are far too chatty or verbose that result in Denial of Service (DOS) of the logging facilities. Thus lets make it available, but repugnant?

In closing. We want Android logging to be the go-to for debugging; but we are in an embedded and limited environment, so we are in a balancing act between overhead, latency, performance and resources.

Patch-set: 2