Update patch set 1
Patch Set 1: Code-Review-1

This isn't core. Move to system/extras.

Don't call it warning, but rather show_images (or even slideshow).
Don't force input forever... I see a reboot in a bag and a battery drain waiting for the second forced input with infinite timeout.

Instead, have it take the timeout as an optional -t argument, then all the other args are just images to show with a timeout or button press in-between.

The state.wait logic is spread out. Contain it within main() and it avoids the unnecessary state stuff: timeouts between images, last image does the infinite wait (but don't because it will drain the battery on reboot in a pocket).

Patch-set: 1
Reviewer: JP Abgrall <1005480@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1