Update patch set 12
Patch Set 12: Code-Review+1 Verified+1

Over-The-Shoulder review resulted in a change where we read into the callers buffer _only_ if it will clear the checks. Otherwise we use lseek to move the file pointer to skip the entries. This solved the imperfect solution of clearing the buffer after read if it failed to pass the checks.

Patch-set: 12
Label: Code-Review=+1
Label: Verified=+1