Update patch set 7
Patch Set 7: Code-Review+1 Verified+1

Look at the other three Related CLs for jni, liblog, logd and logcat. We change the liblog reader API to allow out-of-band events to propagate and be processed by logcat precluding the need to specifically select the events log buffer in order to be notified of data loss incidents.

Does not mean I am happy about the change, this is a prototype idea to see if this addresses Motorola's concerns.

Patch-set: 7
Label: Code-Review=+1
Label: Verified=+1