Update patch set 2
Patch Set 2: Code-Review+2

I will do a cleanup at leisure when builds are not at stake.

I would prefer that we discover a mechanism that defines _all_ the log tags as a deterministic set of defines to be used by the builds. One place to maintain the tag, multiple places just is prone to errors. Any suggestions? We are using the same macro in logd/Android.mk ...

Patch-set: 2
Label: Code-Review=+2