Update patch set 4
Patch Set 4:

(1 comment)

ps4 makes the naming for the system partition match the other partitions and replaces some of the copy & paste with more general code.

the only [deliberate] functional difference here is that now failing to remount the system partition is no longer an error if the system partition doesn't exist. if that's likely to cause trouble in practice, we can add an extra boolean to make /system special.

Patch-set: 4