Update patch set 1
Patch Set 1:

I thought I made it clear, I have _not_ tested this code. It is a prototype to permit discussion. testing, if ever, will take 20 times longer than it took to write this code ;-> My concern was how _many_ capabilities I have allowed this code to support as some may not be relevant or represents security attacks.

I expect that any selinux configuration and filesystem attributes have been set; I have not scaled the entire set of checkmarks required as this is but one step in a long chain. One might question why we are making a wakelock so difficult ;-/

Patch-set: 1