Update patch set 1
Patch Set 1:

(1 comment)

init's setting P(inheritable) (P = process) to get to P'(permitted) is the only field that may have any effect through the exec boundary and _still_ needs help of the F(inheritable) (F = file) flags; and can be set with the F(permitted) flag without. The filesystem is really the only one with control over the executable's capabilities matter. Init can not grant anything unless it is done _after_ the exec at root privs (or if we bypass exec and  wrap-in, a linker running at AID_ROOT that adds them after exec, but before the executable is running)

           P'(permitted) = (P(inheritable) & F(inheritable)) |
                           (F(permitted) & cap_bset)

           P'(effective) = F(effective) ? P'(permitted) : 0

           P'(inheritable) = P(inheritable)

Patch-set: 1