Update patch set 1
Patch Set 1:

maybe long term, but not any time soon. i was mostly concerned that vold seems to be riddled with this error, plus it wasn't clear to me whether fixing BLKGETSIZE callers to use the right size argument is the best fix, or whether we should be moving everyone to BLKGETSIZE64.

but on the assumption that this has been biting you, the vold bugs haven't, and this has been a good enough fix for you so far... let's get it in.

Patch-set: 1