Update patch set 1
Patch Set 1:

Thanks :-) for the +2, and I was ready to 'come over the wall' at you (see following, a valuable treatise nonetheless that can help posterity)

TL;DR

Story:

Ying Wang recanted, and indicated STATIC was prefered as this was part of the ota tools, we do not want to ship the libcutils.so.

In addition, we are using but _one_ module in libcutils.so that is currently an inline in private/android_filesystem_config.h. Once all these CLs have cleared, I will transition from the inline to _actually_ using libcutils. The _weight_ of a static linkage here is _zero_ given that we are replacing an inline, and that inline function is much smaller than the libcutils.so.

Patch-set: 1