Update patch set 1
Patch Set 1:

I believe the burden of proof that this is an unnecessary measure rests with Google and requires a complete device audit.

The statement about comms team comes from a poll of several grandpas on the systems group. I can certainly check current vendor products to prove it (or not), but I do not believe I can check _all_ partner products as to whether they have switched to using radio instead of main for logging their messages. We need to add all those vendors to this CL so they are aware of the change of policy.

Patch-set: 1