Update patch set 1
Patch Set 1:

Re 1+2+4, make_ext4fs shouldn't care if the device is zeroed, it zeroes any data it cares about with fill chunks, if not that is a bug.

Re 3: I'm not particularly concerned about an interrupted fastboot flash command - you should get an error on the host.  An interrupted format or factory reset on the device is more problematic, but this seems like overkill to solve that.  For on-device formatting we could generate a sparse image with a zeroed superblocks, write that, and then manually write the two superblocks.

Patch-set: 1