Update patch set 1
Patch Set 1: Verified+1

this is a simpler version of the other change that listed all the keys used. my guess is that this is probably good enough. if confusion continues, we can always come back to listing the keys, but my current intention is to abandon the other change.

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1