Update patch set 2
Patch Set 2:

We need the spam filtering and reporting to be solid and easy-to-understand.  I looked at the bugs you mention and there's a lot of confusion over what this means.

I'm +2ing this since the code itself looks OK and this will hopefully be a quick fix.  If this doesn't improve things, please seriously consider backing the filtering changes out (again), iterating on them in your local tree more, and running the changes on your own device for a few days (to make sure they're working well in practice) before resubmitting.

Patch-set: 2