Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

Philosophically limiting the logd messages (on a per uid/pid/tid basis though) when reporting spam just makes sense. There was a positive effect on the results when dealing with significant spammers (smith and WifiStateMachine), with a negligible effect on the complicit logs (in the case of WifiStateMachine, trimming other sources of log messages in UID=1000 just because they are also there).

Patch-set: 1
Label: Code-Review=+1
Label: Verified=+1