Update patch set 1
Patch Set 1:

johan.redestig: the problem with your signal handling patch is that it can cause us to use the socketpair before we've set our SELinux label, leading to EPERM, leading to things getting out of sync. switching init over to re-execing itself for labeling is the first step in cleaning up this mess.

Patch-set: 1