Update patch set 1
Patch Set 1:

I've been wanting to add labels to files in rootfs for a while now. I'd like to get rid of all seclabel statements in init.*.rc, and rely exclusively on SELinux auto domain transitions for determining the domain to transfer into.

I like the idea of having init (running in the kernel selinux domain) label rootfs and execute /init to transition to the init selinux domain.

In the future, it also allows other cool things like removing the rootfs from memory, so we execute all code from /system, which could save a bunch of memory.

Patch-set: 1