Update patch set 2
Patch Set 2:

> Why does the android linker consider SIGPIPE a crash worthy of
 > debugging?

[perforce!] history says b/1608633:

I just tracked down a bug in A2DP that was causing the media server to crash without dumping a tombstone.  Since there was no indication of how it failed it was not easy to track down that it was aborting due to a SIGPIPE from a send() call.  Should we be catching SIGPIPE in bionic/linker/debugger.c so problems like this are easier to track down?

Change 130136 by digit@digit-cupcake-mounini on 2009/02/06 16:29:56

        Add SIGPIPE to the list of signals that are going to generate a tombstone
        when a process crashes.
        
        PRESUBMIT=passed
        BUG=1608633
        R=lockwood,swetland,mathias
        CC=android-libc
        APPROVED=lockwood
        DELTA=1  (1 added, 0 deleted, 0 changed)
        OCL=130133

Affected files ...

... //branches/cupcake/android/bionic/linker/debugger.c#2 edit

Patch-set: 2