Update patch set 1
Patch Set 1:

After 279th message is lost the next messages can be sent normally in some delay.
Yes, I see different PID of logd once I set max_dgram_len and kill logd. So, I seem hitting another limit (kernel, logd or emulator). No idea how to dig further. I have this problem filed for a long time but nobody helped. So, this 'band-aid' has been working for my team for a long time and it looks as a reasonable and safe solution/workaround.
I have not tried setting the log frontend. It seems it tried to solve the latency problem, but I need guaranteed delivery. If that patch was not abandoned I would propose to introduce an environment var to set the frontend without application's source modifications. But the additional thread and increased memory usage would be concerns.

Patch-set: 1