Update patch set 1
Patch Set 1:

(5 comments)

I'm not a big fan of "client" either, but I'm not sure what a better name would be. I chose client because it is the program that interfaces to the device side daemon.

I don't like splitting it into "host" and "device" either because then the naming gets confusing if we ever try to run adb on the device (it's been done before, might be done again for more sane reasons).

cli? frontend?

Patch-set: 1
Label: Verified=0