Update patch set 3
Patch Set 3:

(2 comments)

FYI, we are doing our best to follow the Google style rules now. There was no consistent style when we inherited it, which is why it probably stills seems like there's no style to follow :)

When in doubt, there is a clang-format file in the directory.

Patch-set: 3
Label: Verified=0