Update patch set 2
Patch Set 2:

Looking at the daemon code more closely, it looks like the only thing stdout is used for is sending the "OK\n". It then gets redirected to a file. Since it never goes directly to the console, it should be safe to just disable text mode on stdout in the daemon.

Patch-set: 2
Label: Verified=0