Update patch set 1
Patch Set 1:

The point of this commit is to try to cleanup the unix_* vs. adb_* stuff so that future changes (and copy&paste code) are less likely to break Windows (which can't tolerate mismatches between the APIs).

An alternative to this would be to find a way to kill off unix_open(), unix_read(), unix_write(), unix_close() and just embed some logic in adb_open(), adb_read(), adb_write(), adb_close() to work on Windows with well-known file descriptors and C Runtime file descriptors.

I think that might be possible, but it might be too hacky. I figure that it is likely that you guys will rewrite all of this file business, so this is just a quick fix instead.

Patch-set: 1