Update patch set 1
Patch Set 1:

I have evaluated that I do not need this feature in init anymore to solve a problem I had. I was proposing to do the following:

    class_start main
    class_start late_start
    trigger ${logd.syslogd}
. . .
on property:ro.build.type=eng
    setprop logd.syslogd logcatd

on logcatd
    mkdir /data/system/logd 0700 logd logd
    . . .
    start syslogd

To ensure that syslogd started really late so that there was no chance it could hold on to /data on an encrypted device during vold machinations. But a design changed removed the need for this feature being added here in init.

So, do we want it anyways? @enh@ added exec and magically a need for it showed up ;-}

Patch-set: 1