Update patch set 3
Patch Set 3: Code-Review+1 Verified+1

Switched to using libbase. Added check to deal with the fact we could see a year rollover on these logs. Incremented by the time precision so as to prevent the one-log-entry stutter we get on runtime restart. The file match not only checks if it has a trailing dot or nul; but also checks if the first character after the dot is a number (stopped short of checking _all_ characters have to be numbers to keep it in a simple if statement). Allowed -t or -T to override -f check for end of logs to preserve that existing behavior.

Patch-set: 3
Label: Code-Review=+1
Label: Verified=+1