Update patch set 1
Patch Set 1:

Just started discussions with the test team to parse this information.

ObligatoryDisclaimer: Sorry for the churn, explained below.

*I* was the only person doing so, this is so new that no one has likely started. This is the third format I have tried in the past 45 days ... However, the previous and ad-hoc form to quickly address the complaints about the first form of the message (which was all numbers and no names) _did_ get out into preview release and thus lit a fire under me to reset the standard now to my best foot forward or hold my peace.

What is expected is that 'logcat -S' does (or will go further to) produce the tabular information associated with these expire messages in order to get a better bearing on the chatty application(s). The chatty expire messages themselves are there to help people that are triaging determine that there is missing content in the logs; and to provide some temporal data (eg: if they get a burst every 3 minutes of dropped messages, they can spot in the unexpired logs this same burst to find out what is causing the problem). So these messages are for the humans and not those that intend to parse or tabulate the data. logcat -S has more information in its hands than the chatty messages provide.

Patch-set: 1