Update patch set 9
Patch Set 9: Verified+1

This is being held-back until we need to address boot-time delay as a result of logging. The overhead should be small (65ns) but it is added to every log message. One must test the loss in boottime for all others, against the gain in android auto improvement if we turned logging down to 'E' during boot.

Patch-set: 9